+++ src/Generator/Generator.php
@@ -234,8 +234,12 @@ class Generator
         if ($this->options->isEnableAutogeneratePluginPhpAndXmlScript()) {
             $composer_scripts[] = "srag\\LibrariesNamespaceChanger\\GeneratePluginPhpAndXml::generatePluginPhpAndXml";
         }
-        if ($this->options->isEnableUpdatePluginReadmeScript()) {
-            $composer_scripts[] = "srag\LibrariesNamespaceChanger\UpdatePluginReadme::updatePluginReadme";
+        if ($this->options->isEnableAutogeneratePluginReadmeScript()) {
+            $composer_scripts[] = "srag\LibrariesNamespaceChanger\UpdatePluginReadme::generatePluginReadme";
+        } else {
+            if ($this->options->isEnableUpdatePluginReadmeScript()) {
+                $composer_scripts[] = "srag\LibrariesNamespaceChanger\UpdatePluginReadme::updatePluginReadme";
+            }
         }
 
         $ilias_plugin = [